+++ forgero-fabric-compat/src/main/java/com/sigmundgranaas/forgero/fabric/ForgeroCompatClient.java
@@ -1,8 +1,7 @@
 package com.sigmundgranaas.forgero.fabric;
 
-import com.sigmundgranaas.forgero.fabric.ipn.IpnNextCompat;
-import com.sigmundgranaas.forgero.fabric.patchouli.BookDropOnAdvancement;
 import com.sigmundgranaas.forgero.fabric.patchouli.GemUpgradeRecipePage;
+import com.sigmundgranaas.forgero.fabric.patchouli.StateCraftingRecipe;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.loader.api.FabricLoader;
 
@@ -11,8 +10,7 @@ public class ForgeroCompatClient implements ClientModInitializer {
     public void onInitializeClient() {
         if (FabricLoader.getInstance().isModLoaded("patchouli")) {
             GemUpgradeRecipePage.register();
-            BookDropOnAdvancement.registerBookDrop();
+            StateCraftingRecipe.register();
         }
-
     }
 }