+++ forgero-fabric-compat/src/main/java/com/sigmundgranaas/forgero/fabric/ForgeroCompatInitializer.java
@@ -1,6 +1,7 @@
 package com.sigmundgranaas.forgero.fabric;
 
-import com.sigmundgranaas.forgero.fabric.ipn.IpnNextCompat;
+import com.sigmundgranaas.forgero.fabric.patchouli.BookDropOnAdvancement;
+import com.sigmundgranaas.forgero.fabric.patchouli.GuideBookGenerator;
 import com.sigmundgranaas.forgero.fabric.toolstats.ToolStatTagGenerator;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.loader.api.FabricLoader;
@@ -8,12 +9,13 @@ import net.fabricmc.loader.api.FabricLoader;
 public class ForgeroCompatInitializer implements ModInitializer {
     @Override
     public void onInitialize() {
-        if (FabricLoader.getInstance().isModLoaded("inventoryprofilesnext")) {
-            IpnNextCompat.blackListForgero();
-        }
-
         if (FabricLoader.getInstance().isModLoaded("toolstats")) {
             ToolStatTagGenerator.generateTags();
+        }
+
+        if (FabricLoader.getInstance().isModLoaded("patchouli")) {
+            BookDropOnAdvancement.registerBookDrop();
+            GuideBookGenerator.registerGuideBookRecipes();
         }
     }
 }